Create Djdjdj
formatted version with the original HTML source. Let's try introducing an initial paragraph between the H1 and the H2 headings.
Introduction
This file contains a test set of HTML mark-up, as a test of hypertext browsers and an example of the syntax of the tags. See also:

    An arbitrary news article
    The newsgroup on hypertext
    More details about the WWW project. 

That is the end of the list.
Some anchors
Anchors come in two forms: whole nodes or parts of nodes. The line mode browser can't currently (Nov 91) jump to a part of a node: it always jumps to the top.
Leading to whole nodes
Here is an anchor which leads to the VM FIND command. Note the nested highlighting (hp2) within the anchor. If you want to click on this, you will go to the system default page.